requirements.txt
gunicorn
PyMuPDF==1.23.7
Flask==3.1.0
Flask_Cors==5.0.0
googletrans_py==4.0.0
indic_nlp_library==0.92
langdetect==1.0.9
numpy==2.2.5
opencv_python==4.11.0.86
pdf2image==1.17.0
Pillow==11.2.1
pymongo==4.12.1
pytesseract==0.3.13
python-dotenv==1.1.0
torch==2.6.0
transformers==4.50.0
Werkzeug==3.1.3